requirements.txt
flask
requests
Pillow
reportlab
gunicorn
